+++ solutions/devsprint-ana-nogal-1/DeliveryApp/Screens/EmptyView/EmptyView.swift
@@ -4,23 +4,25 @@
 //
 //  Created by Luiza on 26/04/22.
 //
-
 import UIKit
+
+struct EmptyViewConfiguration {
+    var title: String
+    var subtitle: String
+}
 
 final class EmptyView: UIView {
 
     private lazy var titleLabel: UILabel = {
         let label = UILabel()
-        label.text = "Endereço não encontrado"
         label.font = UIFont.systemFont(ofSize: 16, weight: .bold)
         label.textAlignment = .center
         label.translatesAutoresizingMaskIntoConstraints = false
         return label
     }()
-    
+
     private lazy var subtitleLabel: UILabel = {
         let label = UILabel()
-        label.text = "Procure por ruas com número e bairro utilizando o campo de busca"
         label.font = UIFont.systemFont(ofSize: 13, weight: .regular)
         label.textColor = UIColor.lightGray
         label.numberOfLines = 0
@@ -29,7 +31,7 @@ final class EmptyView: UIView {
         label.translatesAutoresizingMaskIntoConstraints = false
         return label
     }()
-    
+
     private lazy var stackView: UIStackView = {
         let stack = UIStackView()
         stack.axis = .vertical
@@ -38,35 +40,38 @@ final class EmptyView: UIView {
         stack.translatesAutoresizingMaskIntoConstraints = false
         return stack
     }()
-    
+
     override init(frame: CGRect) {
         super.init(frame: frame)
         configViews()
         buildViews()
         setupConstraints()
     }
-        
+
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+    
+    func updateView(with configuration: EmptyViewConfiguration) {
+            titleLabel.text = configuration.title
+            subtitleLabel.text = configuration.subtitle
     }
 }
 
 extension EmptyView: ViewConfiguration {
-    internal func configViews() {
+    func configViews() {
         backgroundColor = .white
     }
 
-    internal func buildViews() {
+    func buildViews() {
         addSubview(stackView)
         [titleLabel, subtitleLabel].forEach(stackView.addArrangedSubview)
     }
 
-    internal func setupConstraints() {
+    func setupConstraints() {
         NSLayoutConstraint.activate([
             stackView.centerXAnchor.constraint(equalTo: centerXAnchor),
             stackView.centerYAnchor.constraint(equalTo: centerYAnchor)
         ])
     }
 }
-    
-
